+++ app/ids-progress/index.js
@@ -1,1 +1,0 @@
-import IdsProgress from '../../src/ids-progress/ids-progress';